chore: clear the clippy warnings the clip work introduced
`fragment_bounds` and its test went with the one-fragment-per-GOP shape the
progressive writer replaced — the invariant they guarded, that every fragment
opens on a keyframe, no longer exists to be broken. The rest are two
same-type casts and a `match` on one pattern.

The compile-time relationship between the clip body ceiling and the length
limit moves out of a test and into a `const` block on the relay side, for the
same reason: both sides are constants, so a change that breaks the pairing
should fail the build rather than a test somebody has to remember to run.

The one remaining warning in these files, an elidable lifetime on `covering`,
predates this work.

+++ src/engine/cmaf/fmp4.rs
@@ -2017,8 +2017,8 @@ fn write_progressive_video_trak(
         for &m in &UNITY_MATRIX {
             tkhd.u32(m);
         }
-        tkhd.u32((v.width as u32) << 16);
-        tkhd.u32((v.height as u32) << 16);
+        tkhd.u32(v.width << 16);
+        tkhd.u32(v.height << 16);
     }
     {
         let mut mdia = trak.child(*b"mdia");
